week_3/pyproject.toml
[project]
name = "week-3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mlflow>=2.22.0",
    "pandas>=2.3.0",
    "prefect>=3.4.4",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "scikit-learn~=1.6.1",
]
